+++ packages/core/src/js/index.js
@@ -23,5 +23,6 @@ export { installCommand } from './command.js';
 export { installCalendar } from './calendar.js';
 export { installInputOtp } from './inputotp.js';
 export { installSplitter } from './splitter.js';
+export { installShell } from './shell.js';
 
 export const version = '0.0.0';